+++ docs/src/js/app.js
@@ -1,7 +1,9 @@
 import 'babel-polyfill';
 import React from 'react';
-import Rext from '@/rext';
+import Rext from '~/rext';
 import Viewport from './components/viewport/viewport';
 import GettingStarted from './components/getting-started/getting-started';
+import RextComponent from './components/core-concepts/component';
+import ScreenNavigation from './components/core-concepts/screen-navigation';
 
 Rext.launch(<Viewport />);