+++ docs/src/js/components/viewport/viewport.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { Container, HashRouter } from '@/rext';
+import { Container, HashRouter } from '~/rext';
 import Header from './header';
 import Sidebar from './sidebar';
 import Footer from './footer';