+++ example/js/app/main.js
@@ -1,6 +1,6 @@
 import 'babel-polyfill';
 import React from 'react';
-import Rext from '@/rext';
+import Rext from '~/rext';
 import CommonService from './services/common';
 import Viewport from './components/viewport/viewport';
 import Dashboard from './components/dashboard/dashboard';