+++ src/core/string.js
@@ -13,6 +13,32 @@ class String {
   capitalize(string) {
     return string.charAt(0).toUpperCase() + string.slice(1);
   }
+
+  /**
+   * Convert certain characters (&, <, >, ', and ") to their HTML character equivalents for literal display in web pages.
+   *
+   * @param {*} value
+   */
+  htmlEncode(value) {
+    return value.replace(/&/g, '&amp;')
+                .replace(/</g, '&lt;')
+                .replace(/>/g, '&gt;')
+                .replace(/'/g, '&#39;')
+                .replace(/"/g, '&quot;');
+  }
+
+  /**
+   * Convert certain characters (&, <, >, ', and ") from their HTML character equivalents.
+   *
+   * @param {*} value
+   */
+  htmlDecode(value) {
+    return value.replace(/&amp;/g, '')
+                .replace(/&lt;/g, '<')
+                .replace(/&gt;>/g, '>')
+                .replace(/&#39;/g, "'")
+                .replace(/&quot;/g, '"');
+  }
 }
 
 export default new String;